more PR feedback file validation, expanded testing, and linting
