+++ tasks/change-granule-collection-s3/tests/test-index.js
@@ -4,12 +4,14 @@ const fs = require('fs');
 
 const path = require('path');
 const test = require('ava');
+const keyBy = require('lodash/keyBy');
 const cryptoRandomString = require('crypto-random-string');
 const { s3 } = require('@cumulus/aws-client/services');
 const {
   buildS3Uri,
   recursivelyDeleteS3Bucket,
   putJsonS3Object,
+  s3ObjectExists,
   promiseS3Upload,
   parseS3Uri,
 } = require('@cumulus/aws-client/S3');
@@ -22,8 +24,7 @@ const { isECHO10Filename, isUMMGFilename, metadataObjectFromCMRFile, granulesToC
 
 const { createSnsTopic } = require('@cumulus/aws-client/SNS');
 const { constructCollectionId } = require('../../../packages/message/Collections');
-const { changeGranuleCollectionS3, s3MoveNeeded, s3CopyNeeded, updateCMRData } = require('../dist/src');
-const { keyBy } = require('lodash');
+const { changeGranuleCollectionS3, s3CopyNeeded, updateCMRData } = require('../dist/src');
 
 async function uploadFiles(files) {
   await Promise.all(files.map((file) => {
@@ -297,29 +298,29 @@ function dummyGetGranule(granuleId, t) {
     base_umm_granule: {
       status: 'completed',
       collectionId: 'MOD11A1___006',
-      granuleId: 'MOD11A1.A2017200.h19v04.006.2017201090725',
+      granuleId: 'MOD11A1.A2017200.h19v04.006.2017201090724',
       files: [
         {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090725.hdf',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090725.hdf',
+          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
+          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
           bucket: t.context.protectedBucket,
           type: 'data',
         },
         {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090725_1.jpg',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090725_1.jpg',
+          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
           bucket: t.context.privateBucket,
           type: 'browse',
         },
         {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090725_2.jpg',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090725_2.jpg',
+          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
+          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
           bucket: t.context.publicBucket,
           type: 'browse',
         },
         {
-          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090725.ummg.cmr.json',
-          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090725.ummg.cmr.json',
+          key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.ummg.cmr.json',
+          fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.ummg.cmr.json',
           bucket: t.context.protectedBucket,
           type: 'metadata',
         },
@@ -375,7 +376,6 @@ function buildPayload(t, collection) {
   newPayload.config.testApiClientMethods = t.context.testApiClientMethods;
   return newPayload;
 }
-
 
 test.beforeEach(async (t) => {
   const topicName = randomString();
@@ -433,6 +433,578 @@ test.afterEach.always(async (t) => {
   await recursivelyDeleteS3Bucket(t.context.systemBucket);
 });
 
+test.serial('Should move files to final location with cmr xml file', async (t) => {
+  const payloadPath = path.join(__dirname, 'data', 'payload_cmr_xml.json');
+  t.context.payload = JSON.parse(fs.readFileSync(payloadPath, 'utf8'));
+  const filesToUpload = granulesToFileURIs(
+    t.context.payload.input.granuleIds, t
+  );
+  const collection = { name: 'MOD11A1', version: '002' };
+  const newPayload = buildPayload(t, collection);
+  await uploadFiles(filesToUpload, t.context.bucketMapping);
+  const output = await changeGranuleCollectionS3(newPayload);
+  await validateOutput(t, output);
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.protectedBucket,
+    Key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'jpg/example2/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+  }));
+  const UMM = await metadataObjectFromCMRFile(
+    `s3://${t.context.publicBucket}/example2/2003/` +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml'
+  );
+
+  const CollectionInformation = UMM.Granule.Collection;
+  t.deepEqual(CollectionInformation, { ShortName: 'MOD11A1', VersionId: '002' });
+
+  const onlineResourceUrls = UMM.Granule.OnlineResources.OnlineResource.map(
+    (urlObject) => urlObject.URL
+  );
+  const browseUrls = UMM.Granule.AssociatedBrowseImageUrls.ProviderBrowseUrl.map(
+    (urlObject) => urlObject.URL
+  );
+  const onlineAccessURLs = UMM.Granule.OnlineAccessURLs.OnlineAccessURL.map(
+    (urlObject) => urlObject.URL
+  );
+
+  t.assert(onlineAccessURLs.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.protectedBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.hdf'
+  ));
+  t.assert(browseUrls.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/jpg/example2/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg'
+  ));
+  t.assert(browseUrls.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg'
+  ));
+  t.assert(onlineResourceUrls.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml'
+  ));
+
+  t.assert(onlineAccessURLs.includes(
+    's3://' +
+    `${t.context.protectedBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.hdf'
+  ));
+  t.assert(browseUrls.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/jpg/example2/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg'
+  ));
+  t.assert(browseUrls.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg'
+  ));
+  t.assert(onlineResourceUrls.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml'
+  ));
+});
+
+test.serial('Should move files to final location with cmr umm json file', async (t) => {
+  const payloadPath = path.join(__dirname, 'data', 'payload_cmr_ummg_json.json');
+  t.context.payload = JSON.parse(fs.readFileSync(payloadPath, 'utf8'));
+  const filesToUpload = granulesToFileURIs(
+    t.context.payload.input.granuleIds, t
+  );
+  const collection = { name: 'MOD11A1UMMG', version: '002' };
+  const newPayload = buildPayload(t, collection);
+  await uploadFiles(filesToUpload, t.context.bucketMapping);
+  const output = await changeGranuleCollectionS3(newPayload);
+  await validateOutput(t, output);
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.protectedBucket,
+    Key: 'example2/2016/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'jpg/example2/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'example2/2016/MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'example2/2016/MOD11A1.A2017200.h19v04.006.2017201090724.ummg.cmr.json',
+  }));
+  const UMM = await metadataObjectFromCMRFile(
+    `s3://${t.context.publicBucket}/example2/2016/` +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.ummg.cmr.json'
+  );
+  t.deepEqual(UMM.CollectionReference, { ShortName: 'MOD11A1', Version: '002' });
+  const relatedURLS = UMM.RelatedUrls.map((urlObject) => urlObject.URL);
+
+  t.assert(relatedURLS.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.protectedBucket}` +
+    '/example2/2016/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.hdf'
+  ));
+  t.assert(relatedURLS.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/jpg/example2/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg'
+  ));
+  t.assert(relatedURLS.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/example2/2016/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg'
+  ));
+  t.assert(relatedURLS.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/example2/2016/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.ummg.cmr.json'
+  ));
+
+  t.assert(relatedURLS.includes(
+    's3://' +
+    `${t.context.protectedBucket}` +
+    '/example2/2016/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.hdf'
+  ));
+  t.assert(relatedURLS.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/jpg/example2/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg'
+  ));
+  t.assert(relatedURLS.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/example2/2016/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg'
+  ));
+  t.assert(relatedURLS.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/example2/2016/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.ummg.cmr.json'
+  ));
+});
+
+test.serial('should update cmr data to hold extra urls but remove out-dated urls', async (t) => {
+  const payloadPath = path.join(__dirname, 'data', 'payload_cmr_ummg_json.json');
+  t.context.payload = JSON.parse(fs.readFileSync(payloadPath, 'utf8'));
+  const filesToUpload = granulesToFileURIs(
+    t.context.payload.input.granuleIds, t
+  );
+  const collection = { name: 'MOD11A1UMMG', version: '001' };
+  const newPayload = buildPayload(t, collection);
+  await uploadFiles(filesToUpload, t.context.bucketMapping);
+  const output = await changeGranuleCollectionS3(newPayload);
+  await validateOutput(t, output);
+
+  const UMM = await metadataObjectFromCMRFile(
+    `s3://${t.context.publicBucket}/example2/2016/` +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.ummg.cmr.json'
+  );
+  const URLDescriptions = UMM.RelatedUrls.map((urlObject) => urlObject.Description);
+  // urls that should have been moved are tagged thusly in their description
+  t.false(URLDescriptions.includes('this should be gone by the end'));
+
+  // urls that shouldn't have been changed are tagged thsuly in their description
+  t.assert(URLDescriptions.includes("This should be held onto as it doesn't follow the pattern of tea/s3 url"));
+});
+
+test.serial('handles partially moved files', async (t) => {
+  const payloadPath = path.join(__dirname, 'data', 'payload_cmr_xml.json');
+  t.context.payload = JSON.parse(fs.readFileSync(payloadPath, 'utf8'));
+
+  // a starting granule state that disagrees with the payload as some have already been moved
+  const startingFiles = [
+    {
+      key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
+      fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
+      bucket: t.context.protectedBucket,
+      type: 'data',
+    },
+    {
+      key: 'jpg/example2/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+      fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+      bucket: t.context.publicBucket,
+      type: 'browse',
+    },
+    {
+      key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+      fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+      bucket: t.context.privateBucket,
+      type: 'browse',
+    },
+    {
+      key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
+      fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
+      bucket: t.context.publicBucket,
+      type: 'browse',
+    },
+    {
+      key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+      fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+      bucket: t.context.protectedBucket,
+      type: 'metadata',
+    },
+  ];
+
+  // this is a special case that needs to be in place in massaged form
+  // to be identified as 'not an error'
+  const targetFile = {
+    key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+    fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+    bucket: t.context.publicBucket,
+    type: 'metadata ',
+  };
+
+  const targetXMLBody = fs.readFileSync(
+    path.join(__dirname, 'data', 'target_meta.cmr.xml')
+  ).toString().replaceAll(
+    'replaceme-public', t.context.publicBucket
+  ).replaceAll(
+    'replaceme-protected', t.context.protectedBucket
+  );
+  await promiseS3Upload({
+    params: {
+      Bucket: targetFile.bucket,
+      Key: targetFile.key,
+      Body: targetXMLBody,
+    },
+  });
+  const filesToUpload = startingFiles.map((file) => buildS3Uri(file.bucket, file.key));
+
+  const collection = { name: 'MOD11A1', version: '001' };
+  const newPayload = buildPayload(t, collection);
+
+  await uploadFiles(filesToUpload, t.context.bucketMapping);
+
+  const output = await changeGranuleCollectionS3(newPayload);
+  await validateOutput(t, output);
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.protectedBucket,
+    Key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'jpg/example2/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+  }));
+
+  const UMM = await metadataObjectFromCMRFile(
+    `s3://${t.context.publicBucket}/example2/2003/` +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml'
+  );
+  const onlineAccessURLs = UMM.Granule.OnlineAccessURLs.OnlineAccessURL.map(
+    (urlObject) => urlObject.URL
+  );
+  const onlineResourceUrls = UMM.Granule.OnlineResources.OnlineResource.map(
+    (urlObject) => urlObject.URL
+  );
+  const browseUrls = UMM.Granule.AssociatedBrowseImageUrls.ProviderBrowseUrl.map(
+    (urlObject) => urlObject.URL
+  );
+
+  t.assert(onlineAccessURLs.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.protectedBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.hdf'
+  ));
+  t.assert(browseUrls.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/jpg/example2/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg'
+  ));
+  t.assert(browseUrls.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg'
+  ));
+  t.assert(onlineResourceUrls.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml'
+  ));
+
+  t.assert(onlineAccessURLs.includes(
+    's3://' +
+    `${t.context.protectedBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.hdf'
+  ));
+  t.assert(browseUrls.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/jpg/example2/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg'
+  ));
+  t.assert(browseUrls.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg'
+  ));
+  t.assert(onlineResourceUrls.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml'
+  ));
+});
+
+test.serial('handles files that are pre-moved and misplaced w/r to postgres', async (t) => {
+  const payloadPath = path.join(__dirname, 'data', 'payload_cmr_xml.json');
+  t.context.payload = JSON.parse(fs.readFileSync(payloadPath, 'utf8'));
+  const startingFiles = [
+    {
+      key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
+      bucket: t.context.protectedBucket,
+      type: 'data',
+    },
+    {
+      key: 'jpg/example2/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+      bucket: t.context.publicBucket,
+      type: 'browse',
+    },
+    {
+      key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
+      bucket: t.context.publicBucket,
+      type: 'browse',
+    },
+    {
+      key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+      bucket: t.context.bucketMapping.protected,
+      type: 'metadata',
+    },
+  ];
+
+  const targetFile = {
+    key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+    fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+    bucket: t.context.publicBucket,
+    type: 'metadata ',
+  };
+
+  const targetXMLBody = fs.readFileSync(
+    path.join(__dirname, 'data', 'target_meta.cmr.xml')
+  ).toString().replaceAll(
+    'replaceme-public', t.context.publicBucket
+  ).replaceAll(
+    'replaceme-protected', t.context.protectedBucket
+  );
+  await promiseS3Upload({
+    params: {
+      Bucket: targetFile.bucket,
+      Key: targetFile.key,
+      Body: targetXMLBody,
+    },
+  });
+  const filesToUpload = startingFiles.map((file) => buildS3Uri(file.bucket, file.key));
+  const collection = { name: 'MOD11A1', version: '001' };
+  const newPayload = buildPayload(t, collection);
+
+  await uploadFiles(filesToUpload, t.context.bucketMapping);
+  const output = await changeGranuleCollectionS3(newPayload);
+  await validateOutput(t, output);
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.protectedBucket,
+    Key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'jpg/example2/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+  }));
+
+  const UMM = await metadataObjectFromCMRFile(
+    `s3://${t.context.publicBucket}/example2/2003/` +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml'
+  );
+  const onlineAccessURLs = UMM.Granule.OnlineAccessURLs.OnlineAccessURL.map(
+    (urlObject) => urlObject.URL
+  );
+  const onlineResourceUrls = UMM.Granule.OnlineResources.OnlineResource.map(
+    (urlObject) => urlObject.URL
+  );
+  const browseUrls = UMM.Granule.AssociatedBrowseImageUrls.ProviderBrowseUrl.map(
+    (urlObject) => urlObject.URL
+  );
+
+  t.assert(onlineAccessURLs.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.protectedBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.hdf'
+  ));
+  t.assert(browseUrls.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/jpg/example2/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg'
+  ));
+  t.assert(browseUrls.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg'
+  ));
+  t.assert(onlineResourceUrls.includes(
+    'https://something.api.us-east-1.amazonaws.com/' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml'
+  ));
+
+  t.assert(onlineAccessURLs.includes(
+    's3://' +
+    `${t.context.protectedBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.hdf'
+  ));
+  t.assert(browseUrls.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/jpg/example2/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg'
+  ));
+  t.assert(browseUrls.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg'
+  ));
+  t.assert(onlineResourceUrls.includes(
+    's3://' +
+    `${t.context.publicBucket}` +
+    '/example2/2003/' +
+    'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml'
+  ));
+});
+
+test.serial('handles files that need no move', async (t) => {
+  const payloadPath = path.join(__dirname, 'data', 'payload_cmr_xml.json');
+
+  t.context.payload = JSON.parse(fs.readFileSync(payloadPath, 'utf8'));
+  const filesToUpload = granulesToFileURIs(
+    t.context.payload.input.granuleIds, t
+  );
+  const collection = { name: 'MOD11ANOMOVE', version: '001' };
+  const newPayload = buildPayload(t, collection);
+  newPayload.config.invalidGranuleBehavior = 'error';
+  await uploadFiles(filesToUpload, t.context.bucketMapping);
+  const targetFile = {
+    key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+    fileName: 'MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+    bucket: t.context.bucketMapping.protected,
+    type: 'metadata',
+  };
+  const targetXMLBody = fs.readFileSync(
+    path.join(__dirname, 'data', 'meta.cmr.xml')
+  ).toString().replaceAll(
+    'replaceme-public', t.context.publicBucket
+  ).replaceAll(
+    'replaceme-protected', t.context.protectedBucket
+  );
+  await promiseS3Upload({
+    params: {
+      Bucket: targetFile.bucket,
+      Key: targetFile.key,
+      Body: targetXMLBody,
+    },
+  });
+  const output = await changeGranuleCollectionS3(newPayload);
+  await validateOutput(t, output);
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.protectedBucket,
+    Key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.privateBucket,
+    Key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724_1.jpg',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.publicBucket,
+    Key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724_2.jpg',
+  }));
+  t.assert(await s3ObjectExists({
+    Bucket: t.context.protectedBucket,
+    Key: 'file-staging/subdir/MOD11A1.A2017200.h19v04.006.2017201090724.cmr.xml',
+  }));
+});
+
+test('ignores invalid granules when set to skip', async (t) => {
+  const payloadPath = path.join(__dirname, 'data', 'bad_payload_cmr_xml.json');
+  t.context.payload = JSON.parse(fs.readFileSync(payloadPath, 'utf8'));
+
+  const collection = { name: 'MOD11A1', version: '001' };
+  const newPayload = buildPayload(t, collection);
+  const output = await changeGranuleCollectionS3(newPayload);
+  t.deepEqual(output, {
+    granules: [],
+    oldGranules: [],
+  });
+});
+
+test('errors on invalid granules when set to error', async (t) => {
+  const payloadPath = path.join(__dirname, 'data', 'bad_payload_cmr_xml.json');
+  t.context.payload = JSON.parse(fs.readFileSync(payloadPath, 'utf8'));
+
+  const collection = { name: 'MOD11A1', version: '001' };
+  const newPayload = buildPayload(t, collection);
+  newPayload.config.invalidGranuleBehavior = 'error';
+  try {
+    await changeGranuleCollectionS3(newPayload);
+    t.fail();
+  } catch (error) {
+    t.pass();
+  }
+});
+
 test('s3MoveNeeded checks regular files that arent identical', async (t) => {
   const sourceFile = {
     bucket: t.context.protectedBucket,
@@ -448,7 +1020,7 @@ test('s3MoveNeeded checks regular files that arent identical', async (t) => {
       Key: sourceFile.key,
       Body: 'abcd',
     },
-  })
+  });
 
   t.assert(await s3CopyNeeded(sourceFile, targetFile));
 
@@ -458,9 +1030,9 @@ test('s3MoveNeeded checks regular files that arent identical', async (t) => {
       Key: targetFile.key,
       Body: 'abcd',
     },
-  })
+  });
 
-  t.assert((await s3CopyNeeded(sourceFile, targetFile)) == false);
+  t.assert((await s3CopyNeeded(sourceFile, targetFile)) === false);
 });
 
 test('s3MoveNeeded checks identical files', async (t) => {
@@ -473,7 +1045,7 @@ test('s3MoveNeeded checks identical files', async (t) => {
     key: 'example2/2003/MOD11A1.A2017200.h19v04.006.2017201090724.hdf',
   };
 
-  t.assert((await s3CopyNeeded(sourceFile, targetFile)) == false);
+  t.assert((await s3CopyNeeded(sourceFile, targetFile)) === false);
 });
 
 test('s3MoveNeeded throws if file copy is requested to a location already occupied by different file', async (t) => {
@@ -491,7 +1063,7 @@ test('s3MoveNeeded throws if file copy is requested to a location already occupi
       Key: sourceFile.key,
       Body: 'abcd',
     },
-  })
+  });
 
   await promiseS3Upload({
     params: {
@@ -499,18 +1071,18 @@ test('s3MoveNeeded throws if file copy is requested to a location already occupi
       Key: targetFile.key,
       Body: 'abc',
     },
-  })
+  });
   await t.throwsAsync(
     () => s3CopyNeeded(sourceFile, targetFile),
     { name: 'DuplicateFile' }
   );
 });
 
-test.only('updateCMRData', async (t) => {
+test('updateCMRData', async (t) => {
   const granules = [
     dummyGetGranule('base_xml_granule', t),
     dummyGetGranule('base_umm_granule', t),
-  ]
+  ];
 
   const cmrFiles = granulesToCmrFileObjects(granules, isCMRFile);
   await Promise.all([
@@ -529,7 +1101,7 @@ test.only('updateCMRData', async (t) => {
         Body: isECHO10Filename(cmrFiles[1].key) ? fs.createReadStream('tests/data/meta.cmr.xml') :
           fs.createReadStream('tests/data/ummg-meta.cmr.json'),
       },
-    })
+    }),
   ]);
   const cmrFilesByGranuleid = keyBy(cmrFiles, 'granuleId');
   const CMRObjectsByGranuleId = {};
@@ -551,7 +1123,7 @@ test.only('updateCMRData', async (t) => {
       buckets: {
         internal: {
           name: t.context.stagingBucket,
-          type: "internal",
+          type: 'internal',
         },
         private: {
           name: t.context.privateBucket,
@@ -559,24 +1131,24 @@ test.only('updateCMRData', async (t) => {
         },
         protected: {
           name: t.context.protectedBucket,
-          type: "protected",
+          type: 'protected',
         },
         public: {
           name: t.context.publicBucket,
-          type: "public",
+          type: 'public',
         },
       },
       distribution_endpoint: 'https://something.api.us-east-1.amazonaws.com',
     }
-  )
+  );
   t.assert(getCMRCollectionId(
     updatedCMRData[granules[0].granuleId],
     cmrFilesByGranuleid[granules[0].granuleId].key
-  ) == "abc___003");
+  ) === 'abc___003');
   t.assert(getCMRCollectionId(
     updatedCMRData[granules[1].granuleId],
     cmrFilesByGranuleid[granules[1].granuleId].key
-  ) == "abc___003");
+  ) === 'abc___003');
   const onlineAccessURLs = updatedCMRData[
     'MOD11A1.A2017200.h19v04.006.2017201090724'
   ].Granule.OnlineAccessURLs.OnlineAccessURL.map(
@@ -592,7 +1164,7 @@ test.only('updateCMRData', async (t) => {
   ].Granule.AssociatedBrowseImageUrls.ProviderBrowseUrl.map(
     (urlObject) => urlObject.URL
   );
-  // the following entries are *not in * the original cme metadata file
+  // the following entries are *not in * the original cmr metadata file
   t.assert(onlineAccessURLs.includes(
     'https://something.api.us-east-1.amazonaws.com/' +
     `${t.context.protectedBucket}` +
